Migrate vs detection logic to WindowsOS
